+++ app/todos/page.tsx
@@ -3,12 +3,11 @@
 import { useEffect, useState } from 'react';
 import { useRouter } from 'next/navigation';
 import { api } from '@/lib/api';
-import { getCsrfToken } from '@/lib/csrf';
 import { useUser } from '@/lib/auth';
 import { logout } from '@/lib/auth';
 import LogoutButton from './components/LogoutButton';
 import TaskForm from './components/TaskForm';
-import axios from 'axios';
+import TaskList from './components/TaskList';
 import dayjs from 'dayjs';
 import utc from 'dayjs/plugin/utc';
 import timezone from 'dayjs/plugin/timezone';
@@ -27,15 +26,7 @@ export default function TodosPage() {
   const { user, isLoading, isError } = useUser();
   const [tasks, setTasks] = useState<Task[]>([]);
 
-  const [isToggling, setIsToggling] = useState(false);
-  const [isDeleting, setIsDeleting] = useState(false);
   const [loggingOut, setLoggingOut] = useState(false);
-
-  const [editingTaskId, setEditingTaskId] = useState<number | null>(null);
-  const [editTitle, setEditTitle] = useState('');
-  const [editDueDate, setEditDueDate] = useState('');
-
-  const [isSaving, setIsSaving] = useState(false);
 
   const [errorMessages, setErrorMessages] = useState<string[]>([]);
 
@@ -52,71 +43,6 @@ export default function TodosPage() {
   const fetchTasks = async () => {
     const res = await api.get('/api/tasks');
     setTasks(res.data);
-  };
-
-  const toggleTask = async (task: Task) => {
-    if (isToggling) return;
-    setIsToggling(true);
-    try {
-      await getCsrfToken();
-      await api.put(`/api/tasks/${task.id}`, { is_completed: !task.is_completed });
-      fetchTasks();
-    } catch (e) {
-      console.error(e);
-    } finally {
-      setIsToggling(false);
-    }
-  };
-
-  const deleteTask = async (id: number) => {
-    if (isDeleting) return;
-    setIsDeleting(true);
-    try {
-      await getCsrfToken();
-      await api.delete(`/api/tasks/${id}`);
-      fetchTasks();
-    } catch (e) {
-      console.error(e);
-    } finally {
-      setIsDeleting(false);
-    }
-  };
-
-  const startEdit = (task: Task) => {
-    setEditingTaskId(task.id);
-    setEditTitle(task.title);
-    setEditDueDate(task.due_date || '');
-  };
-
-  const cancelEdit = () => {
-    setEditingTaskId(null);
-    setEditTitle('');
-    setEditDueDate('');
-  };
-
-  const saveTask = async (id: number) => {
-    if (isSaving) return; // 二重送信防止
-    setIsSaving(true);
-
-    try {
-      await getCsrfToken();
-      await api.put(`/api/tasks/${id}`, {
-        title: editTitle,
-        due_date: editDueDate ? dayjs(editDueDate).tz('Asia/Tokyo').format('YYYY-MM-DDTHH:mm:ssZ') : null,
-      });
-      cancelEdit();
-      fetchTasks();
-    } catch (error: unknown) {
-      if (axios.isAxiosError(error) && error.response && error.response.status === 422) {
-        const errors = error.response.data.errors as Record<string, string[]>;
-        const messages = Object.values(errors).flat();
-        setErrorMessages(messages);
-      } else {
-        console.error('Unexpected error:', error);
-      }
-    } finally {
-      setIsSaving(false);
-    }
   };
 
   const handleLogout = async () => {
@@ -151,48 +77,11 @@ export default function TodosPage() {
         </div>
       )}
 
-      <ul className="space-y-2">
-        {tasks.map((task) => (
-          <li key={task.id} className="flex flex-col border p-3">
-            {editingTaskId === task.id ? (
-              <div className="flex flex-col space-y-2">
-                <input
-                  className="border p-2"
-                  value={editTitle}
-                  onChange={(e) => setEditTitle(e.target.value)}
-                />
-                <input
-                  type="datetime-local"
-                  className="border p-2"
-                  value={editDueDate ? dayjs(editDueDate).format('YYYY-MM-DDTHH:mm') : ''}
-                  onChange={(e) => setEditDueDate(e.target.value)}
-                />
-                <div className="flex space-x-2">
-                  <button onClick={() => saveTask(task.id)} disabled={isSaving}>
-                    {isSaving ? '保存中...' : '保存'}
-                  </button>
-                  <button onClick={cancelEdit}>キャンセル</button>
-                </div>
-              </div>
-            ) : (
-              <div className="flex justify-between items-center">
-                <div onClick={() => startEdit(task)} className="cursor-pointer">
-                  <span className={task.is_completed ? 'line-through text-gray-500' : ''}>
-                    {task.title}
-                  </span>
-                  {task.due_date && (
-                    <div className="text-sm text-gray-500">期限: {dayjs(task.due_date).format('YYYY-MM-DD HH:mm')}</div>
-                  )}
-                </div>
-                <div className="flex space-x-2">
-                  <button onClick={() => toggleTask(task)} disabled={isToggling}>完了</button>
-                  <button onClick={() => deleteTask(task.id)} disabled={isDeleting}>削除</button>
-                </div>
-              </div>
-            )}
-          </li>
-        ))}
-      </ul>
+      <TaskList
+        tasks={tasks}
+        setTasks={setTasks}
+        setErrorMessages={setErrorMessages}
+      />
     </div>
   );
 }
